# server/docs/swagger.yaml
basePath: /
definitions:
  handler.ApplicationResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      job:
        $ref: '#/definitions/handler.JobResponse'
      job_id:
        type: integer
      user:
        $ref: '#/definitions/handler.UserResponse'
      user_id:
        type: integer
    type: object
  handler.JobResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      location:
        type: string
      owner_id:
        type: integer
      title:
        type: string
    type: object
  handler.RegisterResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        type: string
    type: object
  handler.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
    type: object
  job.CreateJobInput:
    properties:
      description:
        type: string
      location:
        type: string
      title:
        type: string
    required:
    - description
    - title
    type: object
  user.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  user.LoginOutput:
    properties:
      token:
        type: string
    type: object
  user.RegisterInput:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      role:
        enum:
        - candidate
        - recruiter
        type: string
    required:
    - email
    - password
    - role
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for Recruitment and Selection system
  title: Recruitment API
  version: "1.0"
paths:
  /api/v1/applications/mine:
    get:
      description: Retrieve a list of all jobs the authenticated user has applied
        for
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.ApplicationResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: List my applications
      tags:
      - applications
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: Login Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.LoginOutput'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized / Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login an existing user
      tags:
      - authentication
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user in the system
      parameters:
      - description: Registration Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.RegisterResponse'
        "400":
          description: Invalid input or password too short
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - authentication
  /api/v1/jobs:
    get:
      description: Retrieve a list of all jobs, optionally filtered by a search query
      parameters:
      - description: Search query to filter jobs by title or description
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.JobResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all jobs
      tags:
      - jobs
    post:
      consumes:
      - application/json
      description: Creates a new job posting associated with the authenticated user
      parameters:
      - description: Job Creation Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/job.CreateJobInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.JobResponse'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new job
      tags:
      - jobs
  /api/v1/jobs/{id}/apply:
    post:
      consumes:
      - application/json
      description: Allows the authenticated user to apply to a specific job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.ApplicationResponse'
        "400":
          description: Invalid job ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Job not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Duplicate application
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Cannot apply to own job
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Apply to a Job
      tags:
      - applications
  /api/v1/jobs/mine:
    get:
      description: Retrieve a list of all jobs created by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.JobResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: List my jobs
      tags:
      - jobs
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"